Add feature to change platform colors.
- adds platform prefabs
- collectible item events
- namespaced some scripts
