+++ Assets/Scripts/Characters.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b03ce2ab72584778b1eab5ab7f2a42c2
+timeCreated: 1641366545
