+++ Assets/Scripts/Items.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9dc54777e83245c1bbe0a7c57c51f77e
+timeCreated: 1641366469
