--- Assets/Scripts/PlayerCharacter.cs
@@ -1,14 +1,0 @@
-using System.Collections.Generic;
-using DataModels;
-using UnityEngine;
-
-public class PlayerCharacter : MonoBehaviour
-{
-    public List<ItemData> inventory;
-
-
-    public void CollectItem(ItemData item)
-    {
-        inventory.Add(item);
-    }
-}
